# wire: speed up parsing of headers with arbitrary length connection IDs